+++ .pre-commit-config.yaml
@@ -10,7 +10,7 @@ repos:
       - id: trailing-whitespace
       - id: end-of-file-fixer
       - id: check-docstring-first
-        exclude: ^examples/mosa/lib.py
+        exclude: ^examples/(mosa|dyamond)/lib.py
       - id: check-yaml
 
   - repo: https://github.com/asottile/pyupgrade
