+++ examples/dyamond/run-step1.py
@@ -1,0 +1,5 @@
+"""
+Pre-process DYAMOND files and save Parquet files to TAR archive.
+
+NOTE: requires `pyarrow` to save the GeoDataFrames
+"""
